refactor: revert app dir 💀
