+++ src/pages/_document.tsx
@@ -1,0 +1,19 @@
+import { Html, Head, Main, NextScript } from 'next/document';
+
+export default function Document() {
+    return (
+        <Html lang="en">
+            <Head>
+                <link
+                    rel="shortcut icon"
+                    href="/images/favicon.svg"
+                    type="image/x-icon"
+                />
+            </Head>
+            <body>
+                <Main />
+                <NextScript />
+            </body>
+        </Html>
+    );
+}
