+++ src/pages/index.tsx
@@ -1,0 +1,5 @@
+const Home = () => {
+    return <div>Home</div>;
+};
+
+export default Home;
